dashboard/requirements.txt
flask>=3.0,<4.0
psutil>=5.9
docker>=7.0
psycopg2-binary>=2.9
gunicorn>=22.0
Pillow>=10.0
pillow-heif>=0.14
numpy>=2.0
open-clip-torch>=2.24
torch>=2.0
